% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmeliaMI.R
\name{amelia_impute}
\alias{amelia_impute}
\title{Missing imputation with Amelia.}
\usage{
amelia_impute(data, ID, m, imputed.vars, indep.vars, ...)
}
\arguments{
\item{data}{Input data to be imputed.}

\item{ID}{ID variable in the data.}

\item{m}{Number of times for imputation. Default value: 5. Can be 1.}

\item{rounded}{Logical. If the final imputed values are rounded to the nearest integer or not. Default value: TRUE.}

\item{parallel}{Logical. When TRUE, Amelia runs in a parallel way. Default FALSE.}

\item{verbose}{Logical. Default TRUE.}

\item{imputed.vars}{Variables to be imputed.}

\item{indep.vars}{Independent variables which will be used for missing imputation. If NULL, function will use all the variables except ID in the data for the missing imputattion.}

\item{indep.vars2}{Additional independent variables used for missing imputation. When imputation with \code{indep.vars} fails, additional independent variables will be used for second round imputation. Can be NULL.}

\item{by.var}{A grouping variable. If not NULL, the data will be segmented into many groups, and the missing imputations are done in each group, then the imputed data are combined. If NULL, the data will not be segmented before missing imputation.}
}
\value{
A synthetic data frame with the imputed values. All the imputed values are the average of m imputed datasets.
}
\description{
Conduct missing imputation using Amelia algorithms.
}
\examples{
data.im <- amelia_impute(data = data,
                         ID = "BOOK_ID",
                         imputed.vars = paste0("AT",1:33), 
                         indep.vars   = c(paste0("d",4:16)) 
                         indep.vars2  = c(paste0("d",4:6)),
                         m = 5)

}
