% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMatch.R
\name{append_donor_data}
\alias{append_donor_data}
\title{Append assigned attributes in donor file to the fused data.}
\usage{
append_donor_data(data.fuse = data.fuse, data.don = data.don, by = by,
  vars = vars)
}
\arguments{
\item{data.fuse}{Fused data generated by function \code{stat_match}.}

\item{data.don}{Donor data.}

\item{by}{Variables for table joining. Usually the donor_id plus by.var variable.}

\item{vars}{Variables from donor data which need to be appended.}
}
\value{
A synthetic data frame includes all columns in data.fuse plus the columns of vars from data.don.
}
\description{
Append assigned attributes in donor file to the fused data.
}
\examples{

out <- donor_data_append(out, magtmp, by=c("donor_id", "MAG"), vars=c(paste0("AT",1:10)))

}
