% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{attributes_append}
\alias{attributes_append}
\title{Append attributes of the common variables in both source and target datasets, from the target to the source.}
\usage{
attributes_append(data1 = data1, data2 = data2)
}
\arguments{
\item{data1}{Target dataset.}

\item{data2}{Source dataset.}
}
\value{
The source dataset with attributes of the common variables appended from the target dataset.
}
\description{
Append attributes of the common variables in both source and target datasets, from the target to the source.
}
\examples{
data1 <- read_sav("example1.sav")
data2 <- read_sav("example2.sav")
data1 <- attributes_append(data1 = data1, data2 = data2)
}
