% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCompare.R
\name{combine_comparison_results}
\alias{combine_comparison_results}
\title{Combine several comparison results into one excel spreadsheet.}
\usage{
combine_comparison_results(filename, ...)
}
\arguments{
\item{filename}{Output spreadsheet file name.}

\item{...}{Results from \code{diff_compare}. The function can accept as many comparing results as possible, but at least 2.}
}
\value{
No return value.
}
\description{
Combine several comparison results into one excel spreadsheet.
}
\examples{
res1 <- diff_compare(dat1=data.baseline, dat2=data.primary, wgt1=PERS_WGT, wgt2=PERS_WGT, varlist=varlist, segment="SEGID1")
res2 <- diff_compare(dat1=data.baseline, dat2=data.president.election, wgt1=PERS_WGT, wgt2=PERS_WGT, varlist=varlist, segment="SEGID2")
res3 <- diff_compare(dat1=data.baseline, dat2=data.post.election, wgt1=PERS_WGT, wgt2=PERS_WGT, varlist=varlist, segment="SEGID3")
combine_comparison_results("CompareOutput.xlsx", res1, res2, res3)

}
