% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffCompare.R
\name{diff_compare}
\alias{diff_compare}
\title{Compare two datasets based on relative changes, ztest and ttest.}
\usage{
diff_compare(dat1 = dat1, dat2 = dat2, wgt1 = wgt1, wgt2 = wgt2,
  varlist = fvarlist, segment = segment, ratediff = 0.05)
}
\arguments{
\item{dat1}{dataset 1.}

\item{dat2}{dataset 2.}

\item{wgt1}{weight of dataset1.}

\item{wgt2}{weight of dataset2.}

\item{varlist}{variable list to be compared. All variables must exsit in both datasets.}

\item{segment}{tag to the comparison which is used to identify different comparisons.}

\item{ratediff}{threshold of relative change (percent) for binary variables. All variables will be categoried to dummy variables for comparison at different levels. Default value: 0.05.}
}
\value{
The return list contains \code{comparison results}, and \code{segment identifier}.
}
\description{
Compare two datasets based on relative changes, ztest and ttest.
}
\examples{
res <- diff_compare(dat1=data.baseline, 
                    dat2=data.primary, 
                    wgt1=PERS_WGT,
                    wgt2=PERS_WGT, 
                    varlist=varlist, 
                    segment="SEGID1")

}
