% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{lrm_model}
\alias{lrm_model}
\title{Logistic Regression Model}
\usage{
lrm_model(data = data, DVList = DVList, IDVList = IDVList,
  Included = NULL, all = FALSE, parallel = FALSE)
}
\arguments{
\item{data}{The data. It must contain the variables (columns) that should be used, directly or indirectly, in the modelling procedures. Missing values (NA) are allowed.}

\item{DVList}{Dependent variables list which are needed to be modelled.}

\item{IDVList}{Independent variables list used to model dependent variables.}

\item{Included}{A variable list whose elements are forced to be included in the models, no matter variable selection is conducted or not. Default: NULL.}

\item{all}{Logical. If TRUE, the function will not do variables selection, only model fitting will be done. Default: FALSE.}
}
\value{
A list data structure which contains model results.
}
\description{
Fit binary logistic regression models using MLE or penalized MLE. All the dependent variables are modeled using selected independent variables automatically. Multi-colliearity is removed in the final models using correlation analysis.
}
\examples{

model.Psyco <- lrm_model(data=ModelBase.new, 
                         DVList=DVList,
                         IDVList=PsycoVars)

}
