% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleBalance.R
\name{sample_balance_init}
\alias{sample_balance_init}
\title{Initialize sample balancing process.}
\usage{
sample_balance_init(data = data, targ = targ, out = NULL)
}
\arguments{
\item{data}{Data.}

\item{targ}{An excel spreadsheet which contains targs information.}

\item{out}{Output targs into plain text file. If NULL, no output.}
}
\value{
The targlist.
}
\description{
Initialize sample balancing process.
}
\examples{
data <- read_sav("OnlineWgts.sav")
targ <- read.xlsx("targ.xlsx")

results <- sample_balance_init(data=data, targ=targ, out="targs.txt")
}
