% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{save_model_excel}
\alias{save_model_excel}
\title{Save models to spreadsheet.}
\usage{
save_model_excel(model.out = model.out, out = out)
}
\arguments{
\item{out}{The file name in excel spreadsheet format (.xlsx).}
}
\value{
no return values
}
\description{
Save models information including statistics (R2, AUC), odds, pvalue and vif to excel spreadsheet.
}
\examples{

save_model_excel(model.final, out="Model.xlsx")

}
