% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{score_model_data}
\alias{score_model_data}
\title{Model Deployment: Score model data.}
\usage{
score_model_data(model = model, newdata = newdata, ID = ID,
  cutoff = 0.5, file = file)
}
\arguments{
\item{model}{The model generated by lrm_model.}

\item{newdata}{The modeling data used to build the model.}

\item{ID}{The ID variable in the data.}

\item{cutoff}{The cutoff percent used to create the segment.}

\item{file}{The file name to save the model probabilities and the segments.}
}
\value{
.
}
\description{
Based on the models built, score the modeling base. Probability will be calculated, and the new segments will be generated based on the model scores. Segment has 3 levels. 0: prospects. 1: selected prospects. 2: users. Selected prospected is non users and whose model scores are high than the cutoff value.
}
\examples{

model.final.score <- score_model_data(model = model.final, newdata = ModelBase, ID = "BOOK_ID", cutoff = 0.1, file = "modelscores.xlsx") 

}
